Feat : 툴팁 닫는 로직 훅으로 분리 및 적용
